database/database.py
import os
from dotenv import load_dotenv
from supabase import create_client, Client

load_dotenv()

#Put URL and Key in .env file
url: str = os.getenv("SUPABASE_URL")
key: str = os.getenv("SUPABASE_KEY")

if not url or not key:
    raise ValueError("Supabase credentials missing! Check your .env file.")

supabase: Client = create_client(url, key)