requirements.txt
fastapi
pyswisseph
python-dotenv
supabase
pydantic
uvicorn[standard]
python-dateutil